Remove API that allows for extra deferred runtime asserts during lowering
I want to generate runtime assert nodes during lowering, which means
that I need a finalized list of asserts by the time I start lowering.
This means this runtime assert introduced in
https://github.com/pytorch/pytorch/pull/113839 must go.  Fortunately,
this runtime assert was never exercisable, apparently, and the test
still "passes" without it.  I replace it with a compile time test.  We
can revisit if this assert fails in practice.

Signed-off-by: Edward Z. Yang <ezyang@meta.com>

ghstack-source-id: e93880149f26258c93fcfa79a5fb7c9a87248715
Pull Request resolved: https://github.com/pytorch/pytorch/pull/124864